# spec:/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Roadmap Platform API
  version: 0.1.0
  description: |
    External REST API (BFF prefix `/api`). See `docs/05-api-contracts.md`.

    **Domain rules (P4):** Roadmap items reject initiative/roadmap pairs from different workspaces.
    Strategic themes with a `roadmapId` must target a roadmap in the same workspace. Initiative
    sponsors must belong to the initiative workspace. List endpoints support optional `?workspaceId=`
    for tenant-scoped reads.
servers:
  - url: http://localhost:4000/api
    description: Local API gateway
...
paths:
...
  /phase-segments/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    patch:
      tags: [RoadmapItems]
      summary: Update phase segment
...
  /imports/workbook:
    post:
      tags: [Imports]
      summary: Upload workbook (multipart field `file`)
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                workspaceId:
                  type: string
...